+++ package.json
@@ -120,6 +120,7 @@
     "@types/lodash.uniqby": "^4.7.7",
     "@types/lodash.xor": "^4.5.7",
     "@types/mapbox__geo-viewport": "^0.4.1",
+    "@types/react-dom": "^18.0.0",
     "@types/react-lifecycles-compat": "^3.0.1",
     "@types/react-map-gl": "^6.1.2",
     "@types/react-modal": "^3.13.1",
