--- src/components/map/map-control-tooltip.d.ts
@@ -1,8 +1,0 @@
-import React from 'react';
-
-export type MapControlTooltipProps = {
-  id: string;
-  message: string;
-};
-
-export type MapControlTooltipComponent = React.FunctionComponent<MapControlTooltipProps>;
